+++ src/Media/Media.h
@@ -113,7 +113,7 @@ public:
     void                        setHeight( int height );
 
     unsigned int                getFps() const;
-    void                        setFps( unsigned int fps );
+    void                        setFps( float fps );
 
     const QUuid&                getUuid() const;
 
@@ -148,7 +148,7 @@ protected:
     qint64                      m_length;
     unsigned int                m_width;
     unsigned int                m_height;
-    unsigned int                m_fps;
+    float                       m_fps;
     int*                        m_audioSpectrum;
     audioData                   m_audioData;
     FileType                    m_fileType;
